+++ NEWS.md
@@ -26,6 +26,10 @@
 
 - Line numbers in error messages are incorrect when the error occurs in a child Rmd document (thanks, @cderv @khughitt, #1945).
 
+## MAJOR CHANGES
+
+- Previously `knit_exit()` in a child document would not terminate the processing of its parent document. Now it will (thanks, @davidchall, #1810).
+
 ## MINOR CHANGES
 
 - Vignette engines for R Markdown vignettes no longer check for availability of `pandoc-citeproc` since it has gone since Pandoc 2.11.
@@ -63,7 +67,6 @@
 - The graphics device chunk option, `dev`, now supports a value of `'ragg_png'` which calls the `agg_png()` function from the **ragg** package (thanks, @cpsievert, #1834).
 
 - The argument `error` of `include_graphics()` takes value from the global R option `knitr.graphics.error` by default, e.g., you may set `options(knitr.graphics.error = FALSE)` so `include_graphics()` won't signal an error if the graphics file to be included doesn't exist.
-- Added `propagate` option to `knit_exit()`. This indicates whether the early termination signal should be propagated from child documents to the top-level document. The default value (`FALSE`) retains existing behavior.
 
 - Added a new engine `cc` to compile C++ code via `R CMD SHILIB` like the `c` and `fortran` engines (thanks, @kingaa, #1832).